chore: move links to lib to avoid warnings due to mocks
