+++ apps/trading/pages/client-router.tsx
@@ -3,7 +3,6 @@ import { Navigate, Outlet, useRoutes } from 'react-router-dom';
 import { lazy, Suspense } from 'react';
 import { t } from '@vegaprotocol/i18n';
 import { Loader, Splash } from '@vegaprotocol/ui-toolkit';
-import trimEnd from 'lodash/trimEnd';
 import { LayoutWithSidebar } from '../components/layouts';
 import { LayoutCentered } from '../components/layouts/layout-centered';
 import { Home } from '../client-pages/home';
@@ -19,40 +18,6 @@ import { Transfer } from '../client-pages/transfer';
 // Using dynamic imports is a workaround for this until pennant is published as ESM
 const MarketPage = lazy(() => import('../client-pages/market'));
 const Portfolio = lazy(() => import('../client-pages/portfolio'));
-
-// Make all route paths 'absolute' for easier
-// href creation
-export const Routes = {
-  HOME: '/',
-  MARKETS: '/markets/all',
-  MARKET: '/markets/:marketId',
-  LIQUIDITY: '/liquidity/:marketId',
-  PORTFOLIO: '/portfolio',
-  DISCLAIMER: '/disclaimer',
-  TRANSACT: '/transact',
-  DEPOSIT: '/transact/deposit',
-  WITHDRAW: '/transact/withdraw',
-  TRANSFER: '/transact/transfer',
-} as const;
-
-type ConsoleLinks = {
-  [R in keyof typeof Routes]: (...args: string[]) => string;
-};
-
-export const Links: ConsoleLinks = {
-  HOME: () => Routes.HOME,
-  MARKET: (marketId: string) =>
-    trimEnd(Routes.MARKET.replace(':marketId', marketId)),
-  MARKETS: () => Routes.MARKETS,
-  PORTFOLIO: () => Routes.PORTFOLIO,
-  LIQUIDITY: (marketId: string) =>
-    trimEnd(Routes.LIQUIDITY.replace(':marketId', marketId)),
-  DISCLAIMER: () => Routes.DISCLAIMER,
-  TRANSACT: () => Routes.TRANSACT,
-  DEPOSIT: () => Routes.DEPOSIT,
-  WITHDRAW: () => Routes.WITHDRAW,
-  TRANSFER: () => Routes.TRANSFER,
-};
 
 const NotFound = () => (
   <Splash>
